% Generated by roxygen2 (4.0.1): do not edit by hand
\name{HandleGA}
\alias{HandleGA}
\title{Handle Google Analytics using rga::rga()}
\usage{
HandleGA(start.date = format(Sys.Date() - 8, "\%Y-\%m-\%d"),
  end.date = format(Sys.Date() - 1, "\%Y-\%m-\%d"), domain,
  metrics = "ga:users,ga:sessions,ga:pageviews", dimensions = "ga:date",
  sort = "", filters = "", segment = "", start = 1, max = 10000)
}
\arguments{
\item{start.date}{Start date of the period. %Y-%m-%d}

\item{end.date}{End date of the period. %Y-%m-%d}

\item{domain}{Domain from which extract data. character}

\item{metrics}{Metrics as character (e.g.: "ga:users,ga:sessions,ga:pageviews")}

\item{dimensions}{Dimensions  as character (e.g.: "ga:date")}

\item{sort}{Metrics used to sort (e.g.: "ga:date")}

\item{filters}{Filter (e.g.: "ga:pagePath=~/lx/"")}

\item{segment}{Segment (e.g.: "gaid::-14")}

\item{start}{Start point of the data (usually 1)}

\item{max}{Max record of the data.}
}
\value{
A \code{data.frame}.
}
\description{
Handle Google Analytics using rga::rga()
}
\author{
zmsgnk
}

